requirements.txt
anthropic>=0.18.0
python-libnmap>=0.7.3
pandas>=2.0.0
jinja2>=3.1.0
pytest>=7.4.0
python-dotenv>=1.0.0
requests>=2.31.0
